refactor(gui-v2): remove unnecessary class and attributes
