+++ packages/nocodb/package.json
@@ -181,7 +181,7 @@
     "sha.js": "^2.4.11",
     "slash": "^3.0.0",
     "socket.io": "^4.4.1",
-    "sqlite3": "5.0.0",
+    "sqlite3": "^5.0.0",
     "twilio": "^3.55.1",
     "unique-names-generator": "^4.3.1",
     "uuid": "^8.2.0",
